requirements.txt
asgiref==3.3.4
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.8
decorator==5.0.7
dj-database-url==0.5.0
Django==3.2.2
django-froala-editor==4.0.4
django-heroku==0.3.1
djangorestframework==3.12.4
idna==3.2
importlib-metadata==4.7.1
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
jellyfish==0.8.8
jupyter-client==6.2.0
jupyter-core==4.7.1
lib50==3.0.2
Markdown==3.3.4
markdown2==2.4.0
matplotlib-inline==0.1.2
nest-asyncio==1.5.1
package-name==0.1
paramiko==2.7.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.18
psycopg2==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyNaCl==1.4.0
python-dateutil==2.8.1
python-decouple==3.4
python-utils==2.5.6
pytz==2021.1
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.26.0
six==1.16.0
slugify==0.0.1
sqlparse==0.4.1
submit50==3.1.0
termcolor==1.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
utils==1.0.1
wcwidth==0.2.5
whitenoise==5.3.0
zipp==3.5.0
